+++ spring-cloud-sleuth-autoconfigure/pom.xml
@@ -324,6 +324,11 @@
 			<artifactId>spring-boot-starter-data-mongodb</artifactId>
 			<optional>true</optional>
 		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-rsocket</artifactId>
+			<optional>true</optional>
+		</dependency>
 
 		<!-- Zipkin -->
 		<dependency>
