+++ spring-cloud-sleuth-autoconfigure/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -154,12 +154,6 @@
       "type": "java.lang.Boolean",
       "description": "Enable Spring Cloud Config Server instrumentation.",
       "defaultValue": true
-    },
-    {
-      "name": "spring.sleuth.rsocket.enabled",
-      "type": "java.lang.Boolean",
-      "description": "Enable span information propagation when using RSocket.",
-      "defaultValue": true
     }
   ]
 }
