+++ spring-cloud-sleuth-instrumentation/src/main/java/org/springframework/cloud/sleuth/instrument/rsocket/CompositeMetadataUtils.java
@@ -18,45 +18,24 @@ package org.springframework.cloud.sleuth.instrument.rsocket;
 
 import io.netty.buffer.ByteBuf;
 import io.rsocket.metadata.CompositeMetadata;
-import io.rsocket.metadata.CompositeMetadata.Entry;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Set;
-import reactor.util.annotation.Nullable;
 
-class CompositeMetadataUtils {
+import org.springframework.lang.Nullable;
 
-	static Map<String, ByteBuf> extract(ByteBuf metadata, Set<String> keys) {
-		final CompositeMetadata compositeMetadata = new CompositeMetadata(metadata, false);
+final class CompositeMetadataUtils {
 
-		final Map<String, ByteBuf> map = new HashMap<>();
-		for (Entry entry : compositeMetadata) {
-
-			if (keys.isEmpty()) {
-				return map;
-			}
-
-			final String key = entry.getMimeType();
-			if (keys.contains(key)) {
-				keys.remove(key);
-				map.put(key, entry.getContent());
-			}
-		}
-
-		return map;
+	private CompositeMetadataUtils() {
+		throw new IllegalStateException("Can't instantiate a utility class");
 	}
 
 	@Nullable
 	static ByteBuf extract(ByteBuf metadata, String key) {
 		final CompositeMetadata compositeMetadata = new CompositeMetadata(metadata, false);
-
-		for (Entry entry : compositeMetadata) {
+		for (CompositeMetadata.Entry entry : compositeMetadata) {
 			final String entryKey = entry.getMimeType();
 			if (key.equals(entryKey)) {
 				return entry.getContent();
 			}
 		}
-
 		return null;
 	}
