+++ spring-cloud-sleuth-instrumentation/src/main/java/org/springframework/cloud/sleuth/instrument/rsocket/PayloadUtils.java
@@ -16,6 +16,8 @@
 
 package org.springframework.cloud.sleuth.instrument.rsocket;
 
+import java.util.Set;
+
 import io.netty.buffer.ByteBufAllocator;
 import io.netty.buffer.CompositeByteBuf;
 import io.rsocket.Payload;
@@ -24,31 +26,34 @@ import io.rsocket.metadata.CompositeMetadata.Entry;
 import io.rsocket.metadata.CompositeMetadataCodec;
 import io.rsocket.util.ByteBufPayload;
 import io.rsocket.util.DefaultPayload;
-import java.util.Set;
 
-class PayloadUtils {
+final class PayloadUtils {
+
+	private PayloadUtils() {
+		throw new IllegalStateException("Can't instantiate a utility class");
+	}
 
 	static Payload cleanTracingMetadata(Payload payload, Set<String> fields) {
 		final CompositeMetadata entries = new CompositeMetadata(payload.metadata(), true);
 		final CompositeByteBuf metadata = ByteBufAllocator.DEFAULT.compositeBuffer();
-
 		for (Entry entry : entries) {
 			if (!fields.contains(entry.getMimeType())) {
 				CompositeMetadataCodec.encodeAndAddMetadataWithCompression(metadata, ByteBufAllocator.DEFAULT,
 						entry.getMimeType(), entry.getContent());
 			}
 		}
+		return payload(payload, metadata);
+	}
 
+	private static Payload payload(Payload payload, CompositeByteBuf metadata) {
 		final Payload newPayload;
 		if (payload instanceof ByteBufPayload) {
 			newPayload = ByteBufPayload.create(payload.data().retain(), metadata.retain());
-			payload.release();
 		}
 		else {
 			newPayload = DefaultPayload.create(payload.data().retain(), metadata.retain());
-			payload.release();
 		}
-
+		payload.release();
 		return newPayload;
 	}
