+++ spring-cloud-sleuth-instrumentation/src/main/java/org/springframework/cloud/sleuth/instrument/rsocket/TracingRequesterRSocketProxy.java
@@ -16,7 +16,9 @@
 
 package org.springframework.cloud.sleuth.instrument.rsocket;
 
-import static org.springframework.cloud.sleuth.instrument.rsocket.PayloadUtils.cleanTracingMetadata;
+import java.util.HashSet;
+import java.util.Iterator;
+import java.util.function.Function;
 
 import io.netty.buffer.CompositeByteBuf;
 import io.rsocket.Payload;
@@ -24,42 +26,40 @@ import io.rsocket.RSocket;
 import io.rsocket.metadata.RoutingMetadata;
 import io.rsocket.metadata.WellKnownMimeType;
 import io.rsocket.util.RSocketProxy;
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.function.Function;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.reactivestreams.Publisher;
-import org.springframework.cloud.sleuth.Span;
-import org.springframework.cloud.sleuth.TraceContext;
-import org.springframework.cloud.sleuth.Tracer;
-import org.springframework.cloud.sleuth.propagation.Propagator;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import reactor.util.context.ContextView;
 
-public class RequesterTracingRSocket extends RSocketProxy {
+import org.springframework.cloud.sleuth.Span;
+import org.springframework.cloud.sleuth.TraceContext;
+import org.springframework.cloud.sleuth.Tracer;
+import org.springframework.cloud.sleuth.propagation.Propagator;
 
-	private static final Log log = LogFactory.getLog(RequesterTracingRSocket.class);
+/**
+ * Tracing representation of a {@link RSocketProxy} for the requester.
+ *
+ * @author Marcin Grzejszczak
+ * @author Oleh Dokuka
+ * @since 3.1.0
+ */
+public class TracingRequesterRSocketProxy extends RSocketProxy {
 
-	final Propagator propagator;
+	private static final Log log = LogFactory.getLog(TracingRequesterRSocketProxy.class);
 
-	final Propagator.Setter<CompositeByteBuf> setter;
+	private final Propagator propagator;
 
-	// final MessageSpanCustomizer messageSpanCustomizer;
-	final Tracer tracer;
+	private final Propagator.Setter<CompositeByteBuf> setter;
 
-	//
-	// public ResponderTracingRSocket(RSocket source, Propagator propagator,
-	// Propagator.Setter<ByteBuf> setter, Propagator.Getter<ByteBuf> getter,
-	// MessageSpanCustomizer messageSpanCustomizer, Tracer tracer) {
+	private final Tracer tracer;
 
-	public RequesterTracingRSocket(RSocket source, Propagator propagator, Propagator.Setter<CompositeByteBuf> setter,
-			Tracer tracer) {
+	public TracingRequesterRSocketProxy(RSocket source, Propagator propagator,
+			Propagator.Setter<CompositeByteBuf> setter, Tracer tracer) {
 		super(source);
 		this.propagator = propagator;
 		this.setter = setter;
-		// this.messageSpanCustomizer = messageSpanCustomizer;
 		this.tracer = tracer;
 	}
 
@@ -75,31 +75,29 @@ public class RequesterTracingRSocket extends RSocketProxy {
 
 	<T> Mono<T> setSpan(Function<Payload, Mono<T>> input, Payload payload) {
 		return Mono.deferContextual(contextView -> {
-			Span.Builder spanBuilder = this.tracer.spanBuilder();
-			// TODO: customizing
-			// spanBuilder = this.messageSpanCustomizer.customizeSend(spanBuilder,
-			// message, channel);
-			if (contextView.hasKey(TraceContext.class)) {
-				spanBuilder = spanBuilder.setParent(contextView.get(TraceContext.class));
-			}
-			else if (this.tracer.currentSpan() != null) {
-				// a use case where e.g. rSocketRequest is used outside spring
-				spanBuilder = spanBuilder.setParent(this.tracer.currentSpan().context());
-			}
-
+			Span.Builder spanBuilder = spanBuilder(contextView);
 			final RoutingMetadata routingMetadata = new RoutingMetadata(CompositeMetadataUtils
 					.extract(payload.sliceMetadata(), WellKnownMimeType.MESSAGE_RSOCKET_ROUTING.getString()));
 			final Iterator<String> iterator = routingMetadata.iterator();
-
 			Span span = spanBuilder.kind(Span.Kind.PRODUCER).name(iterator.next()).start();
 			if (log.isDebugEnabled()) {
 				log.debug("Extracted result from context or thread local " + span);
 			}
-			final Payload newPayload = cleanTracingMetadata(payload, new HashSet<>(propagator.fields()));
+			final Payload newPayload = PayloadUtils.cleanTracingMetadata(payload, new HashSet<>(propagator.fields()));
 			this.propagator.inject(span.context(), (CompositeByteBuf) newPayload.metadata(), this.setter);
-
 			return input.apply(newPayload).doOnError(span::error).doFinally(signalType -> span.end());
 		});
+	}
+
+	private Span.Builder spanBuilder(ContextView contextView) {
+		Span.Builder spanBuilder = this.tracer.spanBuilder();
+		if (contextView.hasKey(TraceContext.class)) {
+			spanBuilder = spanBuilder.setParent(contextView.get(TraceContext.class));
+		}
+		else if (this.tracer.currentSpan() != null) {
+			spanBuilder = spanBuilder.setParent(this.tracer.currentSpan().context());
+		}
+		return spanBuilder;
 	}
 
 	@Override
@@ -111,40 +109,25 @@ public class RequesterTracingRSocket extends RSocketProxy {
 	public Flux<Payload> requestChannel(Publisher<Payload> inbound) {
 		return Flux.from(inbound).switchOnFirst((firstSignal, flux) -> {
 			final Payload firstPayload = firstSignal.get();
-
 			if (firstPayload != null) {
 				return setSpan(p -> super.requestChannel(flux.skip(1).startWith(p)), firstPayload,
 						firstSignal.getContextView());
 			}
-
 			return flux;
 		});
 	}
 
 	<T> Flux<Payload> setSpan(Function<Payload, Flux<Payload>> input, Payload payload, ContextView contextView) {
-		Span.Builder spanBuilder = this.tracer.spanBuilder();
-		// TODO: customizing
-		// spanBuilder = this.messageSpanCustomizer.customizeSend(spanBuilder,
-		// message, channel);
-		if (contextView.hasKey(TraceContext.class)) {
-			spanBuilder = spanBuilder.setParent(contextView.get(TraceContext.class));
-		}
-		else if (this.tracer.currentSpan() != null) {
-			// a use case where e.g. rSocketRequest is used outside spring
-			spanBuilder = spanBuilder.setParent(this.tracer.currentSpan().context());
-		}
-
+		Span.Builder spanBuilder = spanBuilder(contextView);
 		final RoutingMetadata routingMetadata = new RoutingMetadata(CompositeMetadataUtils
 				.extract(payload.sliceMetadata(), WellKnownMimeType.MESSAGE_RSOCKET_ROUTING.getString()));
 		final Iterator<String> iterator = routingMetadata.iterator();
-
 		Span span = spanBuilder.kind(Span.Kind.PRODUCER).name(iterator.next()).start();
 		if (log.isDebugEnabled()) {
 			log.debug("Extracted result from context or thread local " + span);
 		}
-		final Payload newPayload = cleanTracingMetadata(payload, new HashSet<>(propagator.fields()));
+		final Payload newPayload = PayloadUtils.cleanTracingMetadata(payload, new HashSet<>(propagator.fields()));
 		this.propagator.inject(span.context(), (CompositeByteBuf) newPayload.metadata(), this.setter);
-
 		return input.apply(newPayload).doOnError(span::error).doFinally(signalType -> span.end());
 	}
