+++ spring-cloud-sleuth-instrumentation/src/main/java/org/springframework/cloud/sleuth/instrument/rsocket/TracingResponderRSocketProxy.java
@@ -16,7 +16,8 @@
 
 package org.springframework.cloud.sleuth.instrument.rsocket;
 
-import static org.springframework.cloud.sleuth.instrument.rsocket.PayloadUtils.cleanTracingMetadata;
+import java.util.HashSet;
+import java.util.Iterator;
 
 import io.netty.buffer.ByteBuf;
 import io.rsocket.Payload;
@@ -25,125 +26,89 @@ import io.rsocket.frame.FrameType;
 import io.rsocket.metadata.RoutingMetadata;
 import io.rsocket.metadata.WellKnownMimeType;
 import io.rsocket.util.RSocketProxy;
-import java.util.HashSet;
-import java.util.Iterator;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.reactivestreams.Publisher;
-import org.springframework.cloud.sleuth.Span;
-import org.springframework.cloud.sleuth.TraceContext;
-import org.springframework.cloud.sleuth.Tracer;
-import org.springframework.cloud.sleuth.propagation.Propagator;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-public class ResponderTracingRSocket extends RSocketProxy {
+import org.springframework.cloud.sleuth.Span;
+import org.springframework.cloud.sleuth.ThreadLocalSpan;
+import org.springframework.cloud.sleuth.TraceContext;
+import org.springframework.cloud.sleuth.Tracer;
+import org.springframework.cloud.sleuth.WithThreadLocalSpan;
+import org.springframework.cloud.sleuth.propagation.Propagator;
 
-	private static final Log log = LogFactory.getLog(ResponderTracingRSocket.class);
+/**
+ * Tracing representation of a {@link RSocketProxy} for the responder.
+ *
+ * @author Marcin Grzejszczak
+ * @author Oleh Dokuka
+ * @since 3.1.0
+ */
+public class TracingResponderRSocketProxy extends RSocketProxy implements WithThreadLocalSpan {
 
-	final Propagator propagator;
+	private static final Log log = LogFactory.getLog(TracingResponderRSocketProxy.class);
 
-	final Propagator.Getter<ByteBuf> getter;
+	private final Propagator propagator;
 
-	// final MessageSpanCustomizer messageSpanCustomizer;
-	final Tracer tracer;
+	private final Propagator.Getter<ByteBuf> getter;
 
-	private final ThreadLocalSpan threadLocalSpan = new ThreadLocalSpan();
+	private final Tracer tracer;
 
-	//
-	// public ResponderTracingRSocket(RSocket source, Propagator propagator,
-	// Propagator.Setter<ByteBuf> setter, Propagator.Getter<ByteBuf> getter,
-	// MessageSpanCustomizer messageSpanCustomizer, Tracer tracer) {
+	private final ThreadLocalSpan threadLocalSpan;
 
-	public ResponderTracingRSocket(RSocket source, Propagator propagator, Propagator.Getter<ByteBuf> getter,
+	public TracingResponderRSocketProxy(RSocket source, Propagator propagator, Propagator.Getter<ByteBuf> getter,
 			Tracer tracer) {
 		super(source);
 		this.propagator = propagator;
 		this.getter = getter;
 		// this.messageSpanCustomizer = messageSpanCustomizer;
 		this.tracer = tracer;
+		this.threadLocalSpan = new ThreadLocalSpan(tracer);
 	}
 
 	@Override
 	public Mono<Void> fireAndForget(Payload payload) {
 		// called on Netty EventLoop
 		// there can't be trace context in thread local here
-		// payload -> metadata -> X-B3-TraceId = a , X-B3-SpanId = b
 		Span consumerSpan = consumerSpan(payload, payload.sliceMetadata(), FrameType.REQUEST_FNF);
 		// create and scope a span for the message processor
 		Span handle = this.tracer.nextSpan(consumerSpan).start();
-		// TODO: Convert Payload to Spring Message?
-		// handle = this.messageSpanCustomizer.customizeHandle(handle, message,
-		// channel).start();
 		if (log.isDebugEnabled()) {
 			log.debug("Created consumer span " + handle);
 		}
 		setSpanInScope(handle);
-
-		// incomingSpan's PARENT - a, b
-		// MessageHeaderPropagatorSetter.removeAnyTraceHeaders(headers,
-		// this.propagator.fields());
-
-		final Payload newPayload = cleanTracingMetadata(payload, new HashSet<>(propagator.fields()));
-
+		final Payload newPayload = PayloadUtils.cleanTracingMetadata(payload, new HashSet<>(propagator.fields()));
 		return super.fireAndForget(newPayload)
-				// TODO: Put TraceContext and Span into reactor context
 				.contextWrite(context -> context.put(Span.class, handle).put(TraceContext.class, handle.context()))
 				.doOnError(this::finishSpan).doOnSuccess(__ -> finishSpan(null)).doOnCancel(() -> finishSpan(null));
 	}
 
 	@Override
 	public Mono<Payload> requestResponse(Payload payload) {
-		// called on Netty EventLoop
-		// there can't be trace context in thread local here
-		// payload -> metadata -> X-B3-TraceId = a , X-B3-SpanId = b
 		Span consumerSpan = consumerSpan(payload, payload.sliceMetadata(), FrameType.REQUEST_RESPONSE);
-		// create and scope a span for the message processor
 		Span handle = this.tracer.nextSpan(consumerSpan).start();
-		// TODO: Convert Payload to Spring Message?
-		// handle = this.messageSpanCustomizer.customizeHandle(handle, message,
-		// channel).start();
 		if (log.isDebugEnabled()) {
 			log.debug("Created consumer span " + handle);
 		}
 		setSpanInScope(handle);
-
-		// incomingSpan's PARENT - a, b
-		// MessageHeaderPropagatorSetter.removeAnyTraceHeaders(headers,
-		// this.propagator.fields());
-
-		final Payload newPayload = cleanTracingMetadata(payload, new HashSet<>(propagator.fields()));
-
+		final Payload newPayload = PayloadUtils.cleanTracingMetadata(payload, new HashSet<>(propagator.fields()));
 		return super.requestResponse(newPayload)
-				// TODO: Put TraceContext and Span into reactor context
 				.contextWrite(context -> context.put(Span.class, handle).put(TraceContext.class, handle.context()))
 				.doOnError(this::finishSpan).doOnSuccess(__ -> finishSpan(null)).doOnCancel(() -> finishSpan(null));
 	}
 
 	@Override
 	public Flux<Payload> requestStream(Payload payload) {
-		// called on Netty EventLoop
-		// there can't be trace context in thread local here
-		// payload -> metadata -> X-B3-TraceId = a , X-B3-SpanId = b
 		Span consumerSpan = consumerSpan(payload, payload.sliceMetadata(), FrameType.REQUEST_STREAM);
-		// create and scope a span for the message processor
 		Span handle = this.tracer.nextSpan(consumerSpan).start();
-		// TODO: Convert Payload to Spring Message?
-		// handle = this.messageSpanCustomizer.customizeHandle(handle, message,
-		// channel).start();
 		if (log.isDebugEnabled()) {
 			log.debug("Created consumer span " + handle);
 		}
 		setSpanInScope(handle);
-
-		// incomingSpan's PARENT - a, b
-		// MessageHeaderPropagatorSetter.removeAnyTraceHeaders(headers,
-		// this.propagator.fields());
-
-		final Payload newPayload = cleanTracingMetadata(payload, new HashSet<>(propagator.fields()));
-
+		final Payload newPayload = PayloadUtils.cleanTracingMetadata(payload, new HashSet<>(propagator.fields()));
 		return super.requestStream(newPayload)
-				// TODO: Put TraceContext and Span into reactor context
 				.contextWrite(context -> context.put(Span.class, handle).put(TraceContext.class, handle.context()))
 				.doOnError(this::finishSpan).doOnComplete(() -> finishSpan(null)).doOnCancel(() -> finishSpan(null));
 	}
@@ -153,28 +118,15 @@ public class ResponderTracingRSocket extends RSocketProxy {
 		return Flux.from(payloads).switchOnFirst((firstSignal, flux) -> {
 			final Payload firstPayload = firstSignal.get();
 			if (firstPayload != null) {
-				// called on Netty EventLoop
-				// there can't be trace context in thread local here
-				// payload -> metadata -> X-B3-TraceId = a , X-B3-SpanId = b
 				Span consumerSpan = consumerSpan(firstPayload, firstPayload.sliceMetadata(), FrameType.REQUEST_CHANNEL);
-				// create and scope a span for the message processor
 				Span handle = this.tracer.nextSpan(consumerSpan).start();
-				// TODO: Convert Payload to Spring Message?
-				// handle = this.messageSpanCustomizer.customizeHandle(handle, message,
-				// channel).start();
 				if (log.isDebugEnabled()) {
 					log.debug("Created consumer span " + handle);
 				}
 				setSpanInScope(handle);
-
-				// incomingSpan's PARENT - a, b
-				// MessageHeaderPropagatorSetter.removeAnyTraceHeaders(headers,
-				// this.propagator.fields());
-
-				final Payload newPayload = cleanTracingMetadata(firstPayload, new HashSet<>(propagator.fields()));
-
+				final Payload newPayload = PayloadUtils.cleanTracingMetadata(firstPayload,
+						new HashSet<>(propagator.fields()));
 				return super.requestChannel(flux.skip(1).startWith(newPayload))
-						// TODO: Put TraceContext and Span into reactor context
 						.contextWrite(
 								context -> context.put(Span.class, handle).put(TraceContext.class, handle.context()))
 						.doOnError(this::finishSpan).doOnComplete(() -> finishSpan(null))
@@ -209,48 +161,9 @@ public class ResponderTracingRSocket extends RSocketProxy {
 		return consumerSpan;
 	}
 
-	// TODO: Copy from SI
-	private void setSpanInScope(Span span) {
-		Tracer.SpanInScope spanInScope = this.tracer.withSpan(span);
-		this.threadLocalSpan.set(new SpanAndScope(span, spanInScope));
-		if (log.isDebugEnabled()) {
-			log.debug("Put span in scope " + span);
-		}
-	}
-
-	// TODO: Copy from SI
-	void finishSpan(Throwable error) {
-		SpanAndScope spanAndScope = getSpanFromThreadLocal();
-		if (spanAndScope == null) {
-			return;
-		}
-		Span span = spanAndScope.span;
-		Tracer.SpanInScope scope = spanAndScope.scope;
-		if (span.isNoop()) {
-			if (log.isDebugEnabled()) {
-				log.debug("Span " + span + " is noop - will stope the scope");
-			}
-			scope.close();
-			return;
-		}
-		if (error != null) { // an error occurred, adding error to span
-			span.error(error);
-		}
-		if (log.isDebugEnabled()) {
-			log.debug("Will finish the and its corresponding scope " + span);
-		}
-		span.end();
-		scope.close();
-	}
-
-	// TODO: Copy from SI
-	private SpanAndScope getSpanFromThreadLocal() {
-		SpanAndScope span = this.threadLocalSpan.get();
-		if (log.isDebugEnabled()) {
-			log.debug("Took span [" + span + "] from thread local");
-		}
-		this.threadLocalSpan.remove();
-		return span;
+	@Override
+	public ThreadLocalSpan getThreadLocalSpan() {
+		return this.threadLocalSpan;
 	}
 
 }
